+++ CHANGELOG.md
@@ -18,6 +18,7 @@ major version hasn't changed.
 - `fiberplane-models`: Add extra types to front matter values, and add a value validation method to
   front matter schema entries.
 - `fiberplane-models`: The `updated_at` and `created_at` fields in the `IntegrationSummary` struct are now optional (#171)
+- Add `PagerDutyWebhooks` to `WorkspaceIntegrationId` (#189)
 
 ## [v1.0.0-beta.14] - 2024-03-07